handleForkAndConcretize: we must add a constraint ourselves if fork is disabled.
Alternative implementation:

remove the check: if (!state->forkDisabled)...
and let Executor::fork() deal with the case when forking is disabled.

This is more complicated, because KLEE's Executor::fork() has non-deterministic
behavior when both outcomes are feasible: it may drop the true state,
which would be pretty hard to handle here.

Signed-off-by: Vitaly Chipounov <vitaly.chipounov@epfl.ch>